+++ ErsatzTV.Core/Interfaces/Repositories/ISearchRepository.cs
@@ -10,6 +10,7 @@ namespace ErsatzTV.Core.Interfaces.Repositories
         Task<List<int>> GetItemIdsToIndex();
         Task<Option<MediaItem>> GetItemToIndex(int id);
         Task<List<string>> GetLanguagesForShow(Show show);
+        Task<List<string>> GetLanguagesForSeason(Season season);
         Task<List<string>> GetLanguagesForArtist(Artist artist);
         Task<List<string>> GetAllLanguageCodes(List<string> mediaCodes);
     }